rest and admin done for nopayers

+++ REST/src/main/java/Entities/NoPayments.java
@@ -7,9 +7,9 @@ import javax.persistence.Id;
 @Entity
 public class NoPayments {
     @Id
-    private int client_id;
+    private long client_id;
 
-    private int number_payments_last_twomonths;
+    private long number_payments_last_twomonths;
 
     public NoPayments() {
     }
@@ -19,7 +19,7 @@ public class NoPayments {
         this.number_payments_last_twomonths = number_payments_last_twomonths;
     }
 
-    public int getClient_id() {
+    public long getClient_id() {
         return client_id;
     }
 
@@ -27,11 +27,11 @@ public class NoPayments {
         this.client_id = client_id;
     }
 
-    public double getNumber_payments_last_twomonths() {
+    public long getNumber_payments_last_twomonths() {
         return number_payments_last_twomonths;
     }
 
-    public void setNumber_payments_last_twomonths(int number_payments_last_twomonths) {
+    public void setNumber_payments_last_twomonths(long number_payments_last_twomonths) {
         this.number_payments_last_twomonths = number_payments_last_twomonths;
     }
 }
